requirements.txt
fastapi
uvicorn
pydantic
openai
pandas
numpy
httpx
python-dotenv
tabulate